// src/core/media/media.events.d.ts
export interface VolumeChange {
	volume: number;
	muted: boolean;
}
// ...
export type VdsMediaEvents = VdsEvents<MediaEvents>;

export class VdsMediaEvent<DetailType> extends VdsCustomEvent<DetailType> {
	static readonly TYPE: keyof VdsMediaEvents;
}
// ...
export class VdsPauseEvent extends VdsMediaEvent<void> {
	static readonly TYPE: 'vds-pause';
	constructor(eventInit?: VdsEventInit<void>);
}

/**
 * Fired when the `paused` property is changed from `true` to `false`, as a result of the `play()`
 * method, or the `autoplay` attribute.
 *
 * @link https://developer.mozilla.org/en-US/docs/Web/API/HTMLMediaElement/play_event
 */
export class VdsPlayEvent extends VdsMediaEvent<void> {
	static readonly TYPE: 'vds-play';
	constructor(eventInit?: VdsEventInit<void>);
}
// ...
export class VdsTimeUpdateEvent extends VdsMediaEvent<number> {
	static readonly TYPE: 'vds-time-update';
	constructor(eventInit: VdsEventInit<number>);
}
// ...
export class VdsVolumeChangeEvent extends VdsMediaEvent<VolumeChange> {
	static readonly TYPE: 'vds-volume-change';
	constructor(eventInit: VdsEventInit<VolumeChange>);
}